add local WAN traffic to table

+++ install.sh
@@ -1,2 +1,3 @@
 #!/bin/sh
-cp ./wrtbmon* /usr/sbin/
+[ -f ./wrtbwmon ] && cp ./wrtbmon* /usr/sbin/ && exit
+wget -O /usr/sbin/wrtbwmon http://raw.githubusercontent.com/pyrovski/wrtbwmon/master/wrtbwmon.sh
